refactor: introduce dropdown to manage search forms
