+++ public/index.html
@@ -46,16 +46,36 @@
             </div>
         </form>
 
-        <form id="blockHeightForm" class="formBlockHeight">
-            <div id="inputBlock">
-                <input type="number" id="blockHeight" name="blockHeight" required>
-                <div id="spinnerHeightForm" class="spinner"></div>
-                <button id="submitHeightButton" type="submit">Submit Height</button>
+        <form id="searchForm">
+            <div id="searchFormTop">
+                <select id="searchType" name="searchType">
+                    <option value="blockHeight">Block Height</option>
+                    <option value="txHash">Transaction Hash</option>
+                    <option value="searchTerm">Search Term</option>
+                </select>
             </div>
-            <div id="fileHelperText">
-                <p id="helperText">Enter a Block Height</p>
+
+            <div id="searchFormBottom">
+                <div class="searchInput" id="blockHeightInput">
+                    <input type="number" id="blockHeight" name="blockHeight" required>
+                    <div class="spinner"></div>
+                </div>
+
+                <div class="searchInput" id="txHashInput">
+                    <input type="text" id="txHash" name="txHash" required>
+                </div>
+
+                <div class="searchInput" id="searchTermInput">
+                    <input type="text" id="searchTerm" name="searchTerm" required>
+                </div>
+
+                <div id="spinnerSearchForm" class="spinner"></div>
+                <button id="submitSearchButton" type="submit">Search</button>
             </div>
         </form>
+
+
+
 
     </div>
 
@@ -64,10 +84,7 @@
     </div>
 
     <script src="/scripts/ses/ses.umd.min.js" charset="utf-8"></script>
-    <script src="./scripts/lockdown.js"></script>
-    <script src="scripts/upload.js"></script>
-    <script src="scripts/submitDateRange.js"></script>
-    <script src="scripts/submitHeight.js"></script>
+    <script type="module" src="./scripts/index.js"></script>
 
 </body>
