+++ public/scripts/index.js
@@ -1,0 +1,4 @@
+import './lockdown.js';
+import './upload.js';
+import './submitDateRange.js';
+import './submitSearch.js';
